auth util created with auth related code removed from app util
